requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
dyu_sis_lib @ file:///Users/nutt1101/nutt/ohin1/code/dyu_sis_lib/dist/dyu_sis_lib-1.0.0b1-py3-none-any.whl#sha256=73f1f1ebbf17d2cc7fb9be173490de372f602285eb3a7ec1c5c5808a5672dd9c
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.7
h11==0.14.0
idna==3.10
motor==3.7.0
pyasn1==0.6.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.11.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
